move stringifyArgv method to flags to avoid circular dep
